Extension: extract shared chat-API clients, manifest for v0.3.0 scope
The ChatGPT/Claude fetch logic moves from the content scripts into
src/lib/chat_apis.ts, parameterized by {origin, init} so the background
worker's upcoming tab-less daily poll runs the exact same client with
credentials: 'include'. Manifest gains the permissions the clipper and
polls need (activeTab, scripting, contextMenus, alarms, notifications,
tabs) and host_permissions for chatgpt.com/chat.openai.com/claude.ai —
no new install-warning surface: content-script matches already warn for
those origins. Renamed "Stash Sync"; v0.3.0.
